requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
PyMySQL==1.1.1
sqlparse==0.5.3
djangorestframework-simplejwt==5.5.0